requirements.txt
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
streamlit
python-dotenv
ollama
fastapi
uvicorn
pydantic
requests
aiohttp
chromadb
beautifulsoup4
lxml
tiktoken
langchain_groq
sentence-transformers
pypdf